added higher order component for user auth

+++ index.js
@@ -17,6 +17,7 @@ app.use(cookieSession({
 }));
 // app.use(express.static(resolve(__dirname, 'client', 'dist')));
 
+
 /******** sign-up endpoint *************/
 
 app.get('/api/cookie-test', (req, res) => {
@@ -24,6 +25,16 @@ app.get('/api/cookie-test', (req, res) => {
         message: 'Session data',
         session: req.session
     });
+});
+
+app.get('/api/is-signed-in', (req, res) => {
+    let auth = false;
+
+    if(req.session.userId){
+        auth = true;
+    }
+
+    res.send({auth});
 });
 
 app.post('/api/sign-up', (req, res) => {
@@ -167,9 +178,10 @@ app.get('/api/products', (req, res) => {
 });
 
 /******* cart endpoint *******/
+
+/******* function getCart(), from /actions/index.js *****************/
 app.get('/api/cart', (req, res) => {
     const userId = req.session.userId;
-    console.log("from cart endpoint, userId:", userId);
 
     // Check if user id, send back error if no id
     if(!userId) {
@@ -219,23 +231,48 @@ app.post('/api/cart', (req, res) => {
 
 /******* cart-meta endpoint *******/
 app.get('/api/cart-meta', (req, res) => {
-    db.query(`SELECT cm.id AS id, c.id AS cart_id, c.customer_id AS customer_id, p.id AS product_id, p.name AS product_name,
-        quantity, p.price AS price, quantity * price AS gross_price FROM \`cart\` AS c, \`products\` AS p, \`cart_meta\` AS cm`, (error, results) => {
+    const userId = req.session.userId;
+
+    // Check if user id, send back error if no id
+    if(!userId) {
+        res.send({
+            success: false,
+            error: "There is no user id"
+        });
+        return;
+    }
+
+    db.query(`SELECT cm.id AS id, c.id AS cart_id, c.customer_id AS customer_id, 
+        p.id AS product_id, p.name AS product_name, quantity, p.price AS price, 
+        quantity * price AS gross_price 
+        FROM \`cart\` AS c, \`products\` AS p, \`cart_meta\` AS cm`, (error, results) => {
         res.send({
             results: results
         });
     });
 });
 
+/*********** function addToCartMeta, from /actions/index.js **********/
 app.post('/api/cart-meta', (req, res) => {
+    const userId = req.session.userId;
+
+    // Check if user id, send back error if no id
+    if(!userId) {
+        res.send({
+            success: false,
+            error: "There is no user id"
+        });
+        return;
+    }
+
     const {product_id, quantity} = req.body;
-    db.query('SELECT * FROM `cart` WHERE `customer_id`=1 AND `status`="incomplete"', (error, result) => {
+    db.query(`SELECT * FROM \`cart\` WHERE customer_id=${userId} AND status="incomplete"`, (error, result) => {
         const inserSql = 'INSERT INTO `cart_meta` (`cart_id`, `customer_id`, `product_id`, `quantity`) VALUES (?, ?, ?, ?)';
 
         if(!result.length){
-            db.query('INSERT INTO `cart` (customer_id, status) VALUES (1, "incomplete")', (error, result) => {
+            db.query(`INSERT INTO \`cart\` (customer_id, status) VALUES (${userId}, "incomplete")`, (error, result) => {
                 const cartId = result.insertId;
-                const inserts = [cartId, 1, product_id, quantity];
+                const inserts = [cartId, userId, product_id, quantity];
                 const sql = mysql.format(inserSql, inserts);
 
                 db.query(sql, (err, result) => {
@@ -248,7 +285,7 @@ app.post('/api/cart-meta', (req, res) => {
         }
 
         const [cart] = result;
-        const inserts = [cart.id, 1, product_id, quantity];
+        const inserts = [cart.id, userId, product_id, quantity];
         const sql = mysql.format(inserSql, inserts);
         db.query(sql, (err, result) => {
             res.send({
@@ -259,10 +296,23 @@ app.post('/api/cart-meta', (req, res) => {
 });
 
 app.delete('/api/cart-meta/product/:id', (req, res) => {
+    const userId = req.session.userId;
+
+    // Check if user id, send back error if no id
+    if(!userId) {
+        res.send({
+            success: false,
+            error: "There is no user id"
+        });
+        return;
+    }
+
     const {id} = req.body;
     if(typeof id !== "number"){
         res.send(422);
+        return;     
     }
+
     db.query(`DELETE FROM \`cart_meta\` WHERE id = ${id}`, (error, results) => {
         res.send({
             results: results
@@ -270,9 +320,22 @@ app.delete('/api/cart-meta/product/:id', (req, res) => {
     });
 });
 
+
+/*************** function updateCartMetaQuantity(), from ../actions/index.js ********/
 app.put('/api/cart-meta/product/:id', (req, res) => {
+    const userId = req.session.userId;
+
+    // Check if user id, send back error if no id
+    if(!userId) {
+        res.send({
+            success: false,
+            error: "There is no user id"
+        });
+        return;
+    }
+
     const {id, quantity} = req.body;
-    db.query(`UPDATE \`cart_meta\` SET \`quantity\` = ${quantity} WHERE \`id\`= ${id}`, (error, results) => {
+    db.query(`UPDATE \`cart_meta\` SET \`quantity\` = ${quantity} WHERE \`id\`= ${id} AND \`customer_id\`=${userId}`, (error, results) => {
         res.send({
             results: results
         });
@@ -290,9 +353,20 @@ app.get('/api/summary', (req, res) => {
 
 app.put('/api/summary/:id', (req, res) => {
     let { id, shipping_method, shipping } = req.body;
-    const user_id = 1;
+    
+    const userId = req.session.userId;
+
+    // Check if user id, send back error if no id
+    if(!userId) {
+        res.send({
+            success: false,
+            error: "There is no user id"
+        });
+        return;
+    }
+
     const query = `SELECT p.name, p.price, i.quantity, c.id AS \`cartId\` FROM cart AS c JOIN products AS p
-        JOIN cart_meta AS i ON c.id=i.cart_id AND i.product_id=p.id WHERE c.status="incomplete" AND c.customer_id=${user_id}`;
+        JOIN cart_meta AS i ON c.id=i.cart_id AND i.product_id=p.id WHERE c.status="incomplete" AND c.customer_id=${userId}`;
 
     db.query(query, (error, results) => {
         if(error){
@@ -314,7 +388,7 @@ app.put('/api/summary/:id', (req, res) => {
             const total = (subTotal * tax) + subTotal + shipping;
             const sql = `UPDATE \`summary\` SET cart_id=?, customer_id=?, total_quantity=?, subtotal=?, tax=?, shipping_method=?, shipping=?, total=?, date=? WHERE \`id\`= ${id}`;
 
-            const inserts = [ cartId, user_id, totalQuantity, subTotal, subTotal * tax, shipping_method, shipping, total, new Date() ];
+            const inserts = [ cartId, userId, totalQuantity, subTotal, subTotal * tax, shipping_method, shipping, total, new Date() ];
             const summaryAdd = mysql.format(sql, inserts);
 
             db.query(summaryAdd, (err, results) => {
@@ -329,10 +403,19 @@ app.put('/api/summary/:id', (req, res) => {
 app.post('/api/summary', (req, res) => {
     db.query(`SELECT * FROM \`summary\``, (error, results) => {
         if(!results.length){
-            const user_id = 1;
+            const userId = req.session.userId;
+
+            // Check if user id, send back error if no id
+            if(!userId) {
+                res.send({
+                    success: false,
+                    error: "There is no user id"
+                });
+                return;
+            }
 
             const query = `SELECT p.name, p.price, i.quantity, c.id AS \`cartId\` FROM cart AS c JOIN products AS p 
-                JOIN cart_meta AS i ON c.id=i.cart_id AND i.product_id=p.id WHERE c.status="incomplete" AND c.customer_id=${user_id}`;
+                JOIN cart_meta AS i ON c.id=i.cart_id AND i.product_id=p.id WHERE c.status="incomplete" AND c.customer_id=${userId}`;
 
             db.query(query, (error, results) => {
                 if(error){
@@ -355,7 +438,7 @@ app.post('/api/summary', (req, res) => {
 
                     const total = (subTotal * tax) + subTotal + shipping;
                     const sql = `INSERT INTO \`summary\` (cart_id, customer_id, total_quantity, subtotal, tax, shipping_method, shipping, total, date) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)`;
-                    const inserts = [ cartId, user_id, totalQuantity, subTotal, subTotal * tax, shipping_method, shipping, total, new Date() ];
+                    const inserts = [ cartId, userId, totalQuantity, subTotal, subTotal * tax, shipping_method, shipping, total, new Date() ];
                     const summaryAdd = mysql.format(sql, inserts);
 
                     db.query(summaryAdd, (err, results) => {
